add paginations for other list pages
