+++ OryAdmin/Components/Pages/Identities/Messages/Index.razor.cs
@@ -1,12 +1,20 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Ory.Kratos.Client.Model;
+using OryAdmin.Extensions;
 using OryAdmin.Services;
+using OryAdmin.Utils;
 
 namespace OryAdmin.Components.Pages.Identities.Messages;
 
 public partial class Index : ComponentBase
 {
     private bool _isLoading = true;
+
+    private PaginationTokens PaginationTokens { get; set; } = null!;
+
+    [SupplyParameterFromQuery(Name = "page_size")]
+    private int PageSize { get; set; }
+    
     private List<KratosMessage>? _messages;
 
     [SupplyParameterFromQuery(Name = "page_token")]
@@ -16,9 +24,12 @@ public partial class Index : ComponentBase
 
     protected override async Task OnInitializedAsync()
     {
-        // TODO use pagination to support a large amount
-        _messages = await ApiService.KratosCourier.ListCourierMessagesAsync();
-
+        if (PageSize == 0) PageSize = 50;
+        
+        var messagesResponse = await ApiService.KratosCourier.ListCourierMessagesWithHttpInfoAsync(PageSize);
+        PaginationTokens = messagesResponse.Headers["Link"].First().PaginationTokens();
+        _messages = messagesResponse.Data;
+        
         _isLoading = false;
     }
 }
