+++ OryAdmin/Components/Pages/OAuth2/Clients/Index.razor.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Ory.Hydra.Client.Model;
+using OryAdmin.Extensions;
 using OryAdmin.Services;
+using OryAdmin.Utils;
 
 namespace OryAdmin.Components.Pages.OAuth2.Clients;
 
@@ -8,14 +10,27 @@ public partial class Index : ComponentBase
 {
     private List<HydraOAuth2Client>? _clients;
     private bool _isLoading = true;
+    
+    [SupplyParameterFromQuery(Name = "page_token")]
+    private string? PageToken { get; set; }
+
+    private PaginationTokens PaginationTokens { get; set; } = null!;
+
+    [SupplyParameterFromQuery(Name = "page_size")]
+    private int PageSize { get; set; }
 
     [Inject] private ApiService ApiService { get; set; } = default!;
 
     protected override async Task OnInitializedAsync()
     {
+        if (PageSize == 0) PageSize = 50;
+        
         try
         {
-            _clients = await ApiService.HydraOAuth2.ListOAuth2ClientsAsync();
+            var clientsResponse = await ApiService.HydraOAuth2
+                .ListOAuth2ClientsWithHttpInfoAsync(PageSize);
+            PaginationTokens = clientsResponse.Headers["Link"].First().PaginationTokens();
+            _clients = clientsResponse.Data;
         }
         catch (Exception)
         {
